[UPD] auth_dynamic_groups: black, isort, prettier

+++ auth_dynamic_groups/models/res_users.py
@@ -1,11 +1,11 @@
 # Copyright 2013-2019 Therp BV <https://therp.nl>
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
-from odoo import models, SUPERUSER_ID
+from odoo import SUPERUSER_ID, models
 from odoo.api import Environment
 
 
 class ResUsers(models.Model):
-    _inherit = 'res.users'
+    _inherit = "res.users"
 
     @classmethod
     def _login(self, db, login, password):
@@ -20,20 +20,17 @@ class ResUsers(models.Model):
     def update_dynamic_groups(self, uid, db):
         with self.pool.cursor() as cr:
             env = Environment(cr, SUPERUSER_ID, {})
-            dynamic_groups = env['res.groups'].search([
-                ('is_dynamic', '=', True)
-            ])
+            dynamic_groups = env["res.groups"].search([("is_dynamic", "=", True)])
             if dynamic_groups:
                 cr.execute(
-                    'delete from res_groups_users_rel '
-                    'where uid=%s and gid in %s',
-                    (uid, tuple(dynamic_groups.ids))
+                    "delete from res_groups_users_rel " "where uid=%s and gid in %s",
+                    (uid, tuple(dynamic_groups.ids)),
                 )
             for dynamic_group in dynamic_groups:
                 if dynamic_group.eval_dynamic_group_condition(uid=uid):
                     cr.execute(
-                        'insert into res_groups_users_rel (uid, gid) values '
-                        '(%s, %s)',
+                        "insert into res_groups_users_rel (uid, gid) values "
+                        "(%s, %s)",
                         (uid, dynamic_group.id),
                     )
-            env['res.users'].invalidate_cache(['groups_id'], [uid])
+            env["res.users"].invalidate_cache(["groups_id"], [uid])
